requirements.txt
setuptools~=46.0.0
python-dotenv~=0.10.3
pymongo~=3.9.0
Faker~=4.1.0